==> 3. Endpoint Configuration.ipynb <==
## Decouple Training & Hosting using Endpoint Configurations

<b> Use Scenario</b><br>
When you want to create an endpoint using Model artifacts trained previously

<b>Note:</b> Let's use the XgBoost Classifier model artifacts we trained previously to create an endpoint here.

##### Imports 

from time import gmtime, strftime
import sagemaker

### Essentials

current_timestamp = strftime("%Y-%m-%d-%H-%M-%S", gmtime())

TRAINING_JOB_NAME = 'classifier-2021-04-15-15-54-02-174' # Copy this from the AWS SageMaker console
MODEL_NAME = f'clf-xgboost-model-{current_timestamp}'
ENDPOINT_CONFIG_NAME = f'clf-xgboost-endpoint-config-{current_timestamp}'
ENDPOINT_NAME = f'clf-xgboost-endpoint-{current_timestamp}'

sagemaker_session = sagemaker.session.Session()

### Create a pointer to your model artifacts using the previously trained job name

sagemaker_session.create_model_from_job(training_job_name=TRAINING_JOB_NAME, 
                                        name=MODEL_NAME)

### Create an Endpoint Config

sagemaker_session.create_endpoint_config(ENDPOINT_CONFIG_NAME, 
                                         model_name=MODEL_NAME, 
                                         initial_instance_count=1, 
                                         instance_type='ml.r5.large')

### Create a new Endpoint 

sagemaker_session.create_endpoint(ENDPOINT_NAME, ENDPOINT_CONFIG_NAME)
